'out_fb' module (support for 24/32-bit Linux framebuffer displays) and fixes for remaining 'byte' coord usage

+++ src/graphics.h
@@ -1,4 +1,4 @@
 #include "types.h"
 
-extern int graphics_drawline(byte x0, byte y0, byte x1, byte y1, RGB *color);
-extern int graphics_drawcircle(byte x0, byte y0, byte radius, RGB *color);
+extern int graphics_drawline(int x0, int y0, int x1, int y1, RGB *color);
+extern int graphics_drawcircle(int x0, int y0, byte radius, RGB *color);
